+++ src/compas_viewer/scene/lineobject.py
@@ -1,39 +1,18 @@
-from typing import List
-from typing import Optional
-from typing import Tuple
-
-from compas.colors import Color
 from compas.geometry import Line
-from compas.geometry import Point
 from compas.scene import GeometryObject
 
+from .sceneobject import DataType
 from .sceneobject import ViewerSceneObject
 
 
 class LineObject(ViewerSceneObject, GeometryObject):
-    """Object for displaying COMPAS line geometry.
-
-    Parameters
-    ----------
-    line : :class:`compas.geometry.Line`
-        The line geometry.
-    **kwargs : Dict, optional
-        Additional options for the :class:`compas_viewer.scene.ViewerSceneObject`.
-
-    """
+    """Viewer scene object for displaying COMPAS :class:`compas.geometry.Line` geometry."""
 
     def __init__(self, line: Line, **kwargs):
         super(LineObject, self).__init__(geometry=line, **kwargs)
-        self._line = line
 
-        self._points_data = self._get_points_data()
-        self._lines_data = self._get_lines_data()
-
-    def _get_points_data(self) -> Optional[Tuple[List[Point], List[Color], List[List[int]]]]:
-        if not self.show_points:
-            return None
-        line = self._line
-        positions = [line.start, line.end]
+    def _read_points_data(self) -> DataType:
+        positions = [self.geometry.start, self.geometry.end]
         colors = [
             self.pointscolor.get(0, self.pointscolor["_default"]),  # type: ignore
             self.pointscolor.get(1, self.pointscolor["_default"]),  # type: ignore
@@ -41,14 +20,11 @@ class LineObject(ViewerSceneObject, GeometryObject):
         elements = [[0], [1]]
         return positions, colors, elements
 
-    def _get_lines_data(self) -> Optional[Tuple[List[Point], List[Color], List[List[int]]]]:
-        line = self._line
-        color = self.linescolor["_default"]
-        positions = [line.start, line.end]
-        colors = [color, color]
+    def _read_lines_data(self) -> DataType:
+        positions = [self.geometry.start, self.geometry.end]
+        colors = [
+            self.pointscolor.get(0, self.pointscolor["_default"]),  # type: ignore
+            self.pointscolor.get(1, self.pointscolor["_default"]),  # type: ignore
+        ]
         elements = [[0, 1]]
         return positions, colors, elements
-
-    @classmethod
-    def create_default(cls) -> Line:
-        return Line([0, 0, 0], [0, 0, 1])