+++ src/compas_viewer/scene/meshobject.py
@@ -1,20 +1,18 @@
-from typing import List
 from typing import Optional
-from typing import Tuple
 
 from compas.colors import Color
 from compas.datastructures import Mesh
-from compas.geometry import Point
 from compas.geometry import centroid_points
 from compas.geometry import is_coplanar
 from compas.scene import MeshObject as BaseMeshObject
 from compas.utilities import pairwise
 
+from .sceneobject import DataType
 from .sceneobject import ViewerSceneObject
 
 
 class MeshObject(ViewerSceneObject, BaseMeshObject):
-    """Object for displaying COMPAS mesh data structures.
+    """Viewer scene object for displaying COMPAS :class:`compas.datastructures.Mesh` geometry.
 
     Parameters
     ----------
@@ -45,8 +43,8 @@ class MeshObject(ViewerSceneObject, BaseMeshObject):
     def __init__(
         self, mesh: Mesh, hide_coplanaredges: Optional[bool] = None, use_vertexcolors: Optional[bool] = None, **kwargs
     ):
-        super(MeshObject, self).__init__(mesh=mesh, **kwargs)
         self._mesh = mesh
+        super(MeshObject, self).__init__(mesh=mesh, **kwargs)
         self.hide_coplanaredges = (
             hide_coplanaredges if hide_coplanaredges is not None else self.config.hide_coplanaredges
         )
@@ -56,14 +54,19 @@ class MeshObject(ViewerSceneObject, BaseMeshObject):
             or self.facescolor.get(vertex, self.facescolor["_default"])  # type: ignore
             for vertex in self._mesh.vertices()
         }
-        self._points_data = self._get_points_data()
-        self._lines_data = self._get_lines_data()
-        self._frontfaces_data = self._get_frontfaces_data()
-        self._backfaces_data = self._get_backfaces_data()
 
-    def _get_points_data(self) -> Optional[Tuple[List[Point], List[Color], List[List[int]]]]:
-        if not self.show_points:
-            return None
+    @property
+    def vertex_xyz(self):
+        """
+        Override :class:`compas.scene.MeshObject`'s property since the viewer does not need to transform the vertices.
+        Transformation is handled by the shader.
+        """
+        if self._vertex_xyz is None:
+            points = self.mesh.vertices_attributes("xyz")  # type: ignore
+            self._vertex_xyz = dict(zip(self.mesh.vertices(), points))  # type: ignore
+        return self._vertex_xyz
+
+    def _read_points_data(self) -> DataType:
         positions = []
         colors = []
         elements = []
@@ -77,9 +80,7 @@ class MeshObject(ViewerSceneObject, BaseMeshObject):
             i += 1
         return positions, colors, elements
 
-    def _get_lines_data(self) -> Optional[Tuple[List[Point], List[Color], List[List[int]]]]:
-        if not self.show_lines:
-            return None
+    def _read_lines_data(self) -> DataType:
         positions = []
         colors = []
         elements = []
@@ -106,9 +107,7 @@ class MeshObject(ViewerSceneObject, BaseMeshObject):
             i += 2
         return positions, colors, elements
 
-    def _get_frontfaces_data(self) -> Optional[Tuple[List[Point], List[Color], List[List[int]]]]:
-        if not self.show_faces:
-            return None
+    def _read_frontfaces_data(self) -> DataType:
         positions = []
         colors = []
         elements = []
@@ -178,9 +177,7 @@ class MeshObject(ViewerSceneObject, BaseMeshObject):
 
         return positions, colors, elements
 
-    def _get_backfaces_data(self) -> Optional[Tuple[List[Point], List[Color], List[List[int]]]]:
-        if not self.show_faces:
-            return None
+    def _read_backfaces_data(self) -> DataType:
         if self.use_vertexcolors:
             self.vertexcolor = {
                 vertex: self._mesh.vertex_attribute(vertex, "color") or Color.grey() for vertex in self._mesh.vertices()